Better handling of invalid JSON components
This commit has some behavioral changes!

An issue regarding the treatment of invalid JSON components was raised this morning. This commit changes the behavior for many things using the component system to instead replace the invalid JSON with text saying "** Invalid JSON Component **". Some things will not do this (e.g. the conversion process).

Here's a list of things that are affected by this change:
- Hover event components (specifically, the show_entity action)
- NBT components
- Entity names
- Beacons
- Enchantment tables
- Scoreboard save data loading